Use mocha & chai for our custom tests instead of nodeunit
test/jsdom/named-properties-tracker has been converted here as a sample
mocha-sugar-free is used for ES6 compatiblity, and also various handy shorthands.

It is not possible to match the jsdom test files using a single glob (test suites and other javascript files are mixed). Because of that, a manifest file is added to each test directory to avoid the need for a big array of paths in the test runners

The old runners & dependencies will be removed in a later commit (after all test cases have been converted)